Move all TripRefine types to one file
