+++ Sources/OJP/Models/OJPv2+StopEvent.swift
@@ -22,19 +22,6 @@ public extension OJPv2 {
         }
     }
 
-    struct TripRefineRequest: Codable, Sendable {
-        public let requestTimestamp: Date
-        public let params: TripRefineParams
-
-        public let tripResult: OJPv2.TripResult
-
-        public enum CodingKeys: String, CodingKey {
-            case requestTimestamp = "siri:RequestTimestamp"
-            case tripResult = "TripResult"
-            case params = "Params"
-        }
-    }
-
     /// [Schema documentation on vdvde.github.io](https://vdvde.github.io/OJP/develop/documentation-tables/ojp.html#type_ojp__StopEventParamStructure)
     struct StopEventParam: Codable, Sendable {
 //        public let modeFilter: ModeFilter
